Исправила замечания, сделала небольшой рефакторинг
