+++ js/main.js
@@ -1,13 +1,17 @@
-//Точка входа
-
 import {data} from './generate-data.js';
-import './render-cards.js';
-import {
-  deactivatePage,
-  activatePage
-} from './form.js';
+import {renderCard} from './render-card.js';
+import {stateToggleElement} from './page-states.js';
 
-deactivatePage();
-activatePage();
-// eslint-disable-next-line no-console
-console.log(data);
+const map = document.querySelector('#map-canvas');
+
+const form = document.querySelector('.ad-form');
+const mapFilterForm = document.querySelector('.map__filters');
+const formItems = form.querySelectorAll('fieldset');
+const mapFlterItems = mapFilterForm.querySelectorAll('select');
+
+
+stateToggleElement(form, formItems);
+stateToggleElement(mapFilterForm, mapFlterItems);
+
+const card = renderCard(data[0]);
+map.appendChild(card);
