+++ js/page-states.js
@@ -1,39 +1,16 @@
-//Работа с формой
+//Переключение состояния формы
 
-const form = document.querySelector('.ad-form');
-const formItems = form.querySelectorAll('fieldset');
-const mapFilterForm = document.querySelector('.map__filters');
-const mapFilter = form.querySelectorAll('fieldset');
+const stateToggleElement = (element, items) => {
+  element.classList.toggle('ad-form--disabled');
 
-const deactivatePage = () => {
-  form.classList.add('ad-form--disabled');
-
-  formItems.forEach((formItem) => {
-    formItem.disabled = true;
-  });
-
-  mapFilterForm.classList.add('ad-form--disabled');
-
-  mapFilter.forEach((formItem) => {
-    formItem.disabled = true;
+  items.forEach((item) => {
+    if(item.disabled === true) {
+      item.disabled = false;
+    } else {
+      item.disabled = true;
+    }
   });
 };
 
-const activatePage = () => {
-  form.classList.remove('ad-form--disabled');
 
-  formItems.forEach((formItem) => {
-    formItem.disabled = false;
-  });
-
-  mapFilterForm.classList.remove('ad-form--disabled');
-
-  mapFilter.forEach((formItem) => {
-    formItem.disabled = false;
-  });
-};
-
-export {
-  deactivatePage,
-  activatePage
-};
+export { stateToggleElement };
